Ensure latency converter uses the right datatype.

+++ src/data/spike_converters/latency_converter.py
@@ -23,7 +23,7 @@ class LatencySpikeConverter(SpikeConverter):
             x_data.shape[2],
             x_data.shape[3],
         )
-        output = np.zeros(out_shape)
+        output = np.zeros(out_shape, dtype=x_data.dtype)
         for i, frame in enumerate(x_data):
             frame = torch.from_numpy(np.moveaxis(frame, 0, -1))
             frame = spikegen.latency(
